delete comments delete unused imports and images separate UserViewModelFactory class
